requirements.txt
PyMySQL==1.0.2
Flask==1.1.2
Pillow==8.1.0
misaka
houdini.py